components and styled components reorganized to different files
